extract notebook outputs preview into separate command
